+++ Changelog.md
@@ -6,6 +6,7 @@
 
 - Add button to remove last segment ([issue #36](https://github.com/sandermvanvliet/RoadCaptain/issues/36))
 - Enable buttons only when they would have an effect (disables reset, save, simulate when on the world selection view) ([issue #39](https://github.com/sandermvanvliet/RoadCaptain/issues/39))
+- Fix a visual bug where the start/end markers of the route would be reversed if the spawn point segment has a reverse direction.
 
 ## 0.6.1.0